Extracted common code in AuthorsFormatter and some other fixes
